Provide default dtype when constructing possibly empty Series in repr tests.

+++ python/cudf/cudf/tests/test_repr.py
@@ -135,8 +135,8 @@ def test_integer_dataframe(x):
 )
 @settings(deadline=None)
 def test_integer_series(x):
-    sr = cudf.Series(x)
-    ps = pd.Series(data=x)
+    sr = cudf.Series(x, dtype=int)
+    ps = pd.Series(data=x, dtype=int)
 
     assert repr(sr) == repr(ps)
 
@@ -144,7 +144,7 @@ def test_integer_series(x):
 @given(st.lists(st.floats()))
 @settings(deadline=None)
 def test_float_dataframe(x):
-    gdf = cudf.DataFrame({"x": cudf.Series(x, nan_as_null=False)})
+    gdf = cudf.DataFrame({"x": cudf.Series(x, dtype=float, nan_as_null=False)})
     pdf = gdf.to_pandas()
     assert repr(gdf) == repr(pdf)
 
@@ -152,8 +152,8 @@ def test_float_dataframe(x):
 @given(st.lists(st.floats()))
 @settings(deadline=None)
 def test_float_series(x):
-    sr = cudf.Series(x, nan_as_null=False)
-    ps = pd.Series(data=x)
+    sr = cudf.Series(x, dtype=float, nan_as_null=False)
+    ps = pd.Series(data=x, dtype=float)
     assert repr(sr) == repr(ps)
